refactor: move 패키지를 piece의 하위 패키지로 이동한다
