+++ src/main/java/chess/domain/piece/move/piece/KingMove.java
@@ -1,16 +1,15 @@
-package chess.domain.move.piece;
+package chess.domain.piece.move.piece;
 
-import chess.domain.move.Direction;
-import chess.domain.move.Location;
 import chess.domain.piece.Position;
+import chess.domain.piece.move.Direction;
+import chess.domain.piece.move.Location;
 
-import static chess.domain.move.piece.Move.MAX_MOVE_COUNT;
+import static chess.domain.piece.move.piece.Move.KING_MAX_MOVE_COUNT;
 
-public class RookMove implements MoveRule {
-
+public final class KingMove implements MoveRule {
     @Override
     public boolean canMove(final Position source, final Position target) {
-        final Location allPositions = Move.getLocation(source, Direction.getFourDirections(), MAX_MOVE_COUNT);
+        final Location allPositions = Move.getLocation(source, Direction.getAllDirections(), KING_MAX_MOVE_COUNT);
         return allPositions.contains(target);
     }
