+++ src/main/java/chess/domain/piece/move/piece/Move.java
@@ -1,8 +1,8 @@
-package chess.domain.move.piece;
+package chess.domain.piece.move.piece;
 
-import chess.domain.move.Direction;
-import chess.domain.move.Location;
 import chess.domain.piece.Position;
+import chess.domain.piece.move.Direction;
+import chess.domain.piece.move.Location;
 
 import java.util.HashSet;
 import java.util.List;
