+++ src/main/java/chess/domain/piece/move/piece/PawnMove.java
@@ -1,12 +1,12 @@
-package chess.domain.move.piece;
+package chess.domain.piece.move.piece;
 
-import chess.domain.move.Direction;
-import chess.domain.move.Location;
 import chess.domain.piece.Position;
+import chess.domain.piece.move.Direction;
+import chess.domain.piece.move.Location;
 
 import java.util.List;
 
-import static chess.domain.move.piece.Move.PAWN_MAX_MOVE_COUNT;
+import static chess.domain.piece.move.piece.Move.PAWN_MAX_MOVE_COUNT;
 
 public class PawnMove implements MoveRule {
     private static final int ATTACK_INDEX = 1;
