+++ src/main/java/chess/domain/piece/move/piece/QueenMove.java
@@ -1,10 +1,10 @@
-package chess.domain.move.piece;
+package chess.domain.piece.move.piece;
 
-import chess.domain.move.Direction;
-import chess.domain.move.Location;
 import chess.domain.piece.Position;
+import chess.domain.piece.move.Direction;
+import chess.domain.piece.move.Location;
 
-import static chess.domain.move.piece.Move.MAX_MOVE_COUNT;
+import static chess.domain.piece.move.piece.Move.MAX_MOVE_COUNT;
 
 public final class QueenMove implements MoveRule {
     @Override
