+++ src/main/java/chess/domain/piece/type/Rook.java
@@ -1,8 +1,8 @@
 package chess.domain.piece.type;
 
 import chess.domain.chess.CampType;
-import chess.domain.move.piece.RookMove;
 import chess.domain.piece.Position;
+import chess.domain.piece.move.piece.RookMove;
 
 public class Rook extends Piece {
