+++ src/test/java/chess/domain/piece/move/LocationTest.java
@@ -1,4 +1,4 @@
-package chess.domain.move;
+package chess.domain.piece.move;
 
 import chess.domain.piece.Position;
 import org.junit.jupiter.api.DisplayName;
